Specifying the collection
I know how to specify the collection now. There are two ways to do it:
1 - within the mongoose model declaration. The tricky part here is that
you have to do it in two places - see twoB.js and oneB1.js.
2 - do this explicitly in the schema declaration -
mongoose.model('carla', OneASchema, 'theCollectionName');
I got this from the ‘model mapped’ section of the first answer at this
link -
http://stackoverflow.com/questions/5794834/how-to-access-a-preexisting-c
ollection-with-mongoose